pyproject.toml
[tool.poetry]
name = "virtbuilder"
version = "0.1.0"
description = ""
authors = ["Panos Mavrogiorgos <pmav99@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
jinja2 = "^2.10"
schema = "^0.6.8"
cleo = "^0.7.2"
"ruamel.yaml" = "^0.15.85"

[tool.poetry.dev-dependencies]
pytest = "^4.0"
pre-commit = "^1.14"
black = {version = "^18.3-alpha.0",allows-prereleases = true}

[tool.poetry.scripts]
virtbuilder = 'virtbuilder.console.application:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
